backend/apps/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
PyMySQL==1.0.0
celery==5.3.0
redis==4.5.0
# pillow==10.0.0  #
python-dotenv==1.0.0
django-redis==5.2.0
drf-yasg==1.21.5